requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch==2.2.2
numpy==1.26.4
pydantic==2.10.3
pillow==10.3.0
aiofiles==23.2.1
python-multipart==0.0.9
